Update README.md
When I use PushModalAsync and run the app on Windows, the "Rate a monkey" text is not shown and the back arrow is not presented. So I'm stuck without ability to navigate away from the rating page. Using PushAsync - and the monkey and ratingState parameters added, it works as described and expected.